Fix regression regarding renamed and edited files
The new extract-submodule.py helper didn't understand how Git prints
files that are renamed and edited in a single commit.  In particular,
it did not understand that these lines can belong to a patch header:

    similarity index 84%
    rename from README
    rename to README.txt

It would treat anything after such a patch as a patch file footer.
This would cause it to miss submodule changes after such a patch.
When this happened, it would print the following message to stderr:

    warning: malformed patch, treated as footer.

This commit applies the changes from commit 1ce8b8e59d04 of the
pdiffdiff repo to extract-submodule.py, thus adding support for
renamed and edited files.  It also adds a test case to t-070.sh to
ensure this works in Guilt.